requirements.txt
langchain==0.3.3
langchain-core==0.3.13
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-chroma==0.1.4
langchain-groq==0.2.0
sqlalchemy==2.0.25
mysql-connector-python
python-dotenv
streamlit
pydantic==2.8.2
ipykernel
